Implement Node SDK config client (#31)
Implements the `PrivyConfig` config client for the Node SDK. This is largely copied over from the private repo.

The first commit in this PR contains files copied verbatim from the existing SDK, and subsequent commits fix it to work in the new codebase.

In order to share the existing `Http` client, I created a new `Session` implementation. The config API uses basic auth instead of token auth though, so I also needed to adjust the Session interface to accommodate this.