Merge Node data and config APIs into single client (#44)
This PR is a light refactor of the Node SDK to merge together the data and config APIs into a single `PrivyClient` class.

To reduce code churn, this was achieved by making the existing `PrivyClient` class into a subclass of `PrivyConfig`. This makes all of the config API methods available on the returned client, and it also automatically makes them visible in the generated docs. (https://github.com/privy-io/privy-docs/pull/21)

With this change, all data API requests from the Node SDK would need to have a single requester ID and roles. This PR uses a hardcoded requester ID of `admin_id` and roles of `['admin']`, but this could change with the upcoming permissions functionality.